Move `PendingBatchesPool` to its own module
